+++ app/seeds/__init__.py
@@ -1,5 +1,7 @@
 from flask.cli import AppGroup
 from .users import seed_users, undo_users
+from .channels import seed_channels, undo_channels
+from .messages import seed_messages, undo_messages
 
 # Creates a seed group to hold our commands
 # So we can type `flask seed --help`
@@ -9,7 +11,17 @@ seed_commands = AppGroup('seed')
 # Creates the `flask seed all` command
 @seed_commands.command('all')
 def seed():
+    print('seeding users and servers')
     seed_users()
+    print('users and servers seeded')
+
+    print('seeding channels')
+    seed_channels()
+    print('channels seeded')
+
+    print('seeding messages')
+    seed_messages()
+    print('messages seeded')
     # Add other seed functions here
